Display finished
Display option finished.

Tested and functional for displaying all songs, and all specific artists (case-sensitive).. WIll ask in class Monday if case-sensitivity is an issue, but the limitations of C make case-sensitivity complicated.

Finished working on this project for the day (9/9/17), and will resume work tomorrow to finish. Logged 7 hours of work today.

Moving on to Edit function